close connection at request shutdown
